refactor: login logic moved to useLogin hook
